Created swing package

+++ src/moneycalculator/model/Money.java
@@ -4,15 +4,15 @@ package moneycalculator.model;
 
 public class Money {
 
-    private final float amount;
+    private final double amount;
     private final Currency currency;
 
-    public Money(float amount, Currency currency) {
+    public Money(double amount, Currency currency) {
         this.amount = amount;
         this.currency = currency;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
